[CCM-9038]: added more filtering condition

+++ ce-nextgen/service/src/main/java/io/harness/ccm/remote/resources/policies/PolicyManagement.java
@@ -11,7 +11,6 @@ import io.harness.ccm.utils.LogAccountIdentifier;
 import io.harness.ng.core.dto.ErrorDTO;
 import io.harness.ng.core.dto.FailureDTO;
 import io.harness.ng.core.dto.ResponseDTO;
-import io.harness.security.annotations.InternalApi;
 import io.harness.security.annotations.NextGenManagerAuth;
 import io.harness.security.annotations.PublicApi;
 
@@ -22,7 +21,6 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -60,13 +58,12 @@ import javax.ws.rs.core.MediaType;
       , @ApiResponse(code = 500, response = ErrorDTO.class, message = "Internal server error")
     })
 
-//@NextGenManagerAuth
-@PublicApi
-public class policyManagement {
+@NextGenManagerAuth
+public class PolicyManagement {
   private final PolicyStoreService policyStoreService;
   private final CCMRbacHelper rbacHelper;
   @Inject
-  public policyManagement(PolicyStoreService policyStoreService, CCMRbacHelper rbacHelper) {
+  public PolicyManagement(PolicyStoreService policyStoreService, CCMRbacHelper rbacHelper) {
     this.policyStoreService = policyStoreService;
     this.rbacHelper = rbacHelper;
   }
@@ -74,8 +71,8 @@ public class policyManagement {
   // Internal API for OOTB policy creation
 
   @POST
-//  @Hidden
-//  @InternalApi
+  //  @Hidden
+  //  @InternalApi
   @Path("add")
   @Consumes(MediaType.APPLICATION_JSON)
   @ApiOperation(value = "Add a new policy internal api", nickname = "addPolicyNameInternal")
@@ -90,7 +87,7 @@ public class policyManagement {
              NGCommonEntityConstants.ACCOUNT_KEY) @AccountIdentifier @NotNull @Valid String accountId,
       @RequestBody(required = true,
           description = "Request body containing Policy store object") @Valid CreatePolicyDTO createPolicyDTO) {
-    //rbacHelper.checkPolicyEditPermission(accountId, null, null);
+    // rbacHelper.checkPolicyEditPermission(accountId, null, null);
     PolicyStore policyStore = createPolicyDTO.getPolicyStore();
     policyStore.setAccountId(accountId);
     policyStoreService.save(policyStore);
@@ -100,8 +97,8 @@ public class policyManagement {
   // Update a policy already made
 
   @PUT
-//  @Hidden
-//  @InternalApi
+  //  @Hidden
+  //  @InternalApi
   @Path("update")
   @Consumes(MediaType.APPLICATION_JSON)
   @ApiOperation(value = "Update a existing OOTB Policy", nickname = "updatePolicy")
@@ -117,7 +114,7 @@ public class policyManagement {
                    NGCommonEntityConstants.ACCOUNT_KEY) @AccountIdentifier @NotNull @Valid String accountId,
       @RequestBody(required = true,
           description = "Request body containing ceViewFolder object") @Valid CreatePolicyDTO createPolicyDTO) {
-    //rbacHelper.checkPolicyEditPermission(accountId, null, null);
+    rbacHelper.checkPolicyEditPermission(accountId, null, null);
     PolicyStore policyStore = createPolicyDTO.getPolicyStore();
     policyStore.toDTO();
     policyStore.setAccountId(accountId);
@@ -128,8 +125,8 @@ public class policyManagement {
   // Internal API for deletion of OOTB policies
 
   @DELETE
-//  @Hidden
-//  @InternalApi
+  //  @Hidden
+  //  @InternalApi
   @Path("{policyId}")
   @Timed
   @ExceptionMetered
@@ -148,7 +145,7 @@ public class policyManagement {
              NGCommonEntityConstants.ACCOUNT_KEY) @AccountIdentifier @NotNull @Valid String accountId,
       @PathParam("policyId") @Parameter(
           required = true, description = "Unique identifier for the policy") @NotNull @Valid String uuid) {
-   //rbacHelper.checkPolicyDeletePermission(accountId, null, null);
+    rbacHelper.checkPolicyDeletePermission(accountId, null, null);
     boolean result = policyStoreService.delete(accountId, uuid);
     return ResponseDTO.newResponse(result);
   }
@@ -170,7 +167,7 @@ public class policyManagement {
                  NGCommonEntityConstants.ACCOUNT_KEY) @AccountIdentifier @NotNull @Valid String accountId,
       @RequestBody(
           required = true, description = "Request body containing ceViewFolder object") @Valid ListDTO listDTO) {
- //rbacHelper.checkPolicyViewPermission(accountId, null, null);
+    rbacHelper.checkPolicyViewPermission(accountId, null, null);
     QueryFeild query = listDTO.getQueryFeild();
     List<PolicyStore> Policies = new ArrayList<>();
     query.setAccountId(accountId);
@@ -178,13 +175,17 @@ public class policyManagement {
     String orgIdentifier = query.getOrgIdentifier();
     String projectIdentifier = query.getProjectIdentifier();
     String cloudProvider = query.getCloudProvider();
-    String isStablePolicy= query.getIsStablePolicy();
-    String isOOTBPolicy= query.getIsOOTBPolicy();
+    String isStablePolicy = query.getIsStablePolicy();
+    String isOOTBPolicy = query.getIsOOTBPolicy();
     String resource = query.getResource();
     String tags = query.getTags();
     if (uuid != null) {
       PolicyStore policyStore = policyStoreService.listid(accountId, uuid);
       Policies.add(policyStore);
+      return ResponseDTO.newResponse(Policies);
+    }
+    if (isStablePolicy != null) {
+      Policies = policyStoreService.findByStability(isStablePolicy, accountId);
       return ResponseDTO.newResponse(Policies);
     }
     if (resource != null && tags == null) {
